RFID Based Fuel Theft Prevention
The purpose of this report is to introduce a new technology that will prevent the fuel theft in any organization. Right now, there are so many companies which have gas station inside their premises to provide fuel to their own vehicles. Sometimes the workers or out-siders fill-in their own vehicles without the authority. To prevent this situation, we are going to develop a system which will help the organization to stop the fuel theft and they can save their fuel.

Features Implemented
1. Authentication and Authorization
2. Security
3. Data Generation
4. Remote Monitoring